% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{malawi_palettes}
\alias{malawi_palettes}
\title{Malawi Themed Color Palettes}
\format{
A list of character vectors containing hex color codes.
}
\usage{
malawi_palettes
}
\description{
A collection of professionally designed color palettes optimized for
Malawi maps and data visualization. Palettes are carefully chosen to be
colorblind-friendly and print-ready.
}
\keyword{internal}
