% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_level_0}
\alias{mw_level_0}
\title{Malawi Administrative Boundaries - Level 0 (Country)}
\format{
An sf object with 1 feature and 3 fields:
\describe{
\item{ADM0_EN}{Country name (Malawi)}
\item{ADM0_PCODE}{Country ISO code (MWI)}
\item{geometry}{MULTIPOLYGON geometry for the national boundary}
}
}
\source{
Malawi Spatial Data Platform (MASDAP) / Department of Surveys
}
\usage{
mw_level_0
}
\description{
National boundary of Malawi at administrative level 0. This represents the
outermost border of the country, including major lakes.
}
\keyword{datasets}
