% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mw_level_1}
\alias{mw_level_1}
\title{Malawi Administrative Boundaries - Level 1 (Regions)}
\format{
An sf object with 3 features and 7 fields:
\describe{
\item{ADM1_EN}{Region name (Northern, Central, Southern)}
\item{ADM1_PCODE}{Region code (MW001, MW002, MW003)}
\item{ADM0_EN}{Country name}
\item{ADM0_PCODE}{Country code}
\item{Shape_Leng}{Perimeter length}
\item{Shape_Area}{Area in square meters}
\item{geometry}{MULTIPOLYGON geometry for regional boundaries}
}
}
\source{
Malawi Spatial Data Platform (MASDAP) / Department of Surveys
}
\usage{
mw_level_1
}
\description{
Administrative regions of Malawi (Northern, Central, Southern) at level 1.
These are the three main administrative regions.
}
\keyword{datasets}
