% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_join.R
\name{mw_suggest_matches}
\alias{mw_suggest_matches}
\title{Suggest Close Malawi Name Matches}
\usage{
mw_suggest_matches(x, candidates, n = 3)
}
\arguments{
\item{x}{Character. Name to match.}

\item{candidates}{Character vector of valid names.}

\item{n}{Number of suggestions.}
}
\value{
Character vector of suggested names.
}
\description{
Suggest Close Malawi Name Matches
}
\examples{
mw_suggest_matches("Lilongwee", mw_districts())
}
