% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{view_mw_palettes}
\alias{view_mw_palettes}
\title{Visualize Malawi Palettes}
\usage{
view_mw_palettes(n = NULL, ncol = 3)
}
\arguments{
\item{n}{Number of colors to show for each palette. Default: NULL (all colors).}

\item{ncol}{Number of columns in plot. Default: 3.}
}
\value{
A ggplot2 object showing palette swatches.
}
\description{
Display all available palettes for visual inspection.
}
\examples{
\donttest{
# Show all palettes
view_mw_palettes()

# Show 5 colors from each palette
view_mw_palettes(n = 5)
}

}
