% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw_districts.R
\name{mw_districts_simple}
\alias{mw_districts_simple}
\title{Simple version that always works}
\usage{
mw_districts_simple(region = NULL)
}
\arguments{
\item{region}{Optional region filter}
}
\value{
Character vector of district names
}
\description{
Simple version that always works
}
